fix: bug where documents without inherits became invisible in the inheritance graph
