+++ integration/schema-language-server/language-server/src/main/java/ai/vespa/schemals/index/DocumentInheritanceGraph.java
@@ -8,6 +8,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 /*
  * This class is responsible for managing inheritance relationships among documents
@@ -31,6 +32,16 @@ public class DocumentInheritanceGraph {
         }
 
         documentParents.remove(fileURI);
+    }
+
+    public void createNodeIfNotExists(String fileURI) {
+        if (!documentParents.containsKey(fileURI)) {
+            documentParents.put(fileURI, new ArrayList<>());
+        }
+
+        if (!documentChildren.containsKey(fileURI)) {
+            documentChildren.put(fileURI, new ArrayList<>());
+        }
     }
 
     /*
@@ -127,7 +138,7 @@ public class DocumentInheritanceGraph {
 
         for (String parentURI : documentChildren.keySet()) {
             logger.println(getFileName(parentURI) + " has children:");
-            for (String childURI : getChildren(parentURI)) {
+            for (String childURI : getValidChildren(parentURI)) {
                 logger.println("    " + getFileName(childURI));
             }
         }
@@ -164,7 +175,7 @@ public class DocumentInheritanceGraph {
 
         result.add(fileURI);
 
-        for (String childURI : getChildren(fileURI)) {
+        for (String childURI : getValidChildren(fileURI)) {
             getAllDocumentDescendantURIsImpl(childURI, result, visited);
         }
 
@@ -174,27 +185,10 @@ public class DocumentInheritanceGraph {
         return documentParents.containsKey(fileURI) && documentChildren.containsKey(fileURI);
     }
 
-    private void createNodeIfNotExists(String fileURI) {
-        if (!documentParents.containsKey(fileURI)) {
-            documentParents.put(fileURI, new ArrayList<>());
-        }
-
-        if (!documentChildren.containsKey(fileURI)) {
-            documentChildren.put(fileURI, new ArrayList<>());
-        }
-    }
-
-    private List<String> getChildren(String fileURI) {
-        List<String> childrenList = documentChildren.getOrDefault(fileURI, new ArrayList<>());
-
-        List<String> correctList = new ArrayList<>();
-        for (String childURI : childrenList) {
-            if (documentParents.get(childURI).contains(fileURI)) {
-                correctList.add(childURI);
-            }
-        }
-
-        documentChildren.put(fileURI, correctList);
-        return correctList;
+    private List<String> getValidChildren(String fileURI) {
+        return documentChildren.getOrDefault(fileURI, new ArrayList<>())
+                               .stream()
+                               .filter(childURI -> documentParents.get(childURI).contains(fileURI))
+                               .collect(Collectors.toList());
     }
 }
